set default dop=1
