+++ example/session/src/main/java/org/apache/iotdb/DataMigrationExample.java
@@ -24,6 +24,8 @@ import org.apache.iotdb.rpc.IoTDBConnectionException;
 import org.apache.iotdb.rpc.StatementExecutionException;
 import org.apache.iotdb.session.pool.SessionPool;
 import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
+import org.apache.iotdb.tsfile.read.common.Path;
+import org.apache.iotdb.tsfile.utils.Binary;
 import org.apache.iotdb.tsfile.write.record.Tablet;
 import org.apache.iotdb.tsfile.write.schema.MeasurementSchema;
 
@@ -49,60 +51,55 @@ public class DataMigrationExample {
   // used to write data into the destination IoTDB
   private static SessionPool writerPool;
   // concurrent thread of loading timeseries data
+  private static int concurrency = 5;
+
   public static void main(String[] args)
       throws IoTDBConnectionException, StatementExecutionException, ExecutionException,
           InterruptedException {
 
-    long start = System.currentTimeMillis();
-    String path = "root.**";
-    args = new String[] {"127.0.0.1", "6667", "root", "127.0.0.1", "6668", "root", "1"};
-    ExecutorService executorService = null;
-    if (args == null || args.length != 7) {
-      System.out.println(
-          "Correct input parameters are required. For example, java -jar xxx.jar sourceIP sourcePort sourcePassword destIP destPort destPassword concurrency(ThreadNumber = 2*concurrency+1)");
-      System.exit(1);
-    }
-    try {
-      int concurrency = Integer.parseInt(args[6]);
-      executorService = Executors.newFixedThreadPool(2 * concurrency + 1);
-      readerPool =
-          new SessionPool(args[0], Integer.parseInt(args[1]), "root", args[2], 2 * concurrency + 1);
-      writerPool =
-          new SessionPool(args[3], Integer.parseInt(args[4]), "root", args[5], 2 * concurrency + 1);
-    } catch (Exception e) {
-      System.out.println(
-          "Correct input parameters are required. For example, java -jar xxx.jar sourceIP sourcePort sourcePassword destIP destPort destPassword concurrency(ThreadNumber = 2*concurrency+1)");
-      System.exit(1);
+    ExecutorService executorService = Executors.newFixedThreadPool(2 * concurrency + 1);
+
+    String path = "root";
+
+    if (args.length != 0) {
+      path = args[0];
     }
 
-    SessionDataSetWrapper deviceDataSet = readerPool.executeQueryStatement("count devices " + path);
-    DataIterator deviceIter = deviceDataSet.iterator();
+    readerPool = new SessionPool("127.0.0.1", 6667, "root", "root", concurrency);
+    writerPool = new SessionPool("127.0.0.1", 6668, "root", "root", concurrency);
+
+    SessionDataSetWrapper schemaDataSet =
+        readerPool.executeQueryStatement("count timeseries " + path);
+    DataIterator schemaIter = schemaDataSet.iterator();
     int total;
-    if (deviceIter.next()) {
-      total = deviceIter.getInt(1);
-      System.out.println("Total devices: " + total);
+    if (schemaIter.next()) {
+      total = schemaIter.getInt(1);
+      System.out.println("Total timeseries: " + total);
     } else {
-      System.out.println("Can not get devices schema");
+      System.out.println("Can not get timeseries schema");
       System.exit(1);
     }
-    readerPool.closeResultSet(deviceDataSet);
+    readerPool.closeResultSet(schemaDataSet);
 
-    deviceDataSet = readerPool.executeQueryStatement("show devices " + path);
-    deviceIter = deviceDataSet.iterator();
+    schemaDataSet = readerPool.executeQueryStatement("show timeseries " + path);
+    schemaIter = schemaDataSet.iterator();
 
     List<Future> futureList = new ArrayList<>();
     int count = 0;
-    while (deviceIter.next()) {
+    while (schemaIter.next()) {
       count++;
-      Future future = executorService.submit(new LoadThread(count, deviceIter.getString("Device")));
+      Path currentPath = new Path(schemaIter.getString("Timeseries"), true);
+      Future future =
+          executorService.submit(
+              new LoadThread(
+                  count, currentPath, TSDataType.valueOf(schemaIter.getString("DataType"))));
       futureList.add(future);
     }
-    readerPool.closeResultSet(deviceDataSet);
+    readerPool.closeResultSet(schemaDataSet);
 
     for (Future future : futureList) {
       future.get();
     }
-    System.out.println("Total cost: " + (System.currentTimeMillis() - start) + "ms");
     executorService.shutdown();
 
     readerPool.close();
@@ -112,114 +109,77 @@ public class DataMigrationExample {
   static class LoadThread implements Callable<Void> {
 
     String device;
+    String measurement;
+    Path series;
+    TSDataType dataType;
     Tablet tablet;
     int i;
 
-    public LoadThread(int i, String device) {
+    public LoadThread(int i, Path series, TSDataType dataType) {
       this.i = i;
-      this.device = device;
+      this.device = series.getDevice();
+      this.measurement = series.getMeasurement();
+      this.dataType = dataType;
+      this.series = series;
     }
 
     @Override
     public Void call() {
+
+      List<MeasurementSchema> schemaList = new ArrayList<>();
+      schemaList.add(new MeasurementSchema(measurement, dataType));
+      tablet = new Tablet(device, schemaList, 300000);
       SessionDataSetWrapper dataSet = null;
-      long startTime = System.nanoTime();
+
       try {
-        dataSet = readerPool.executeQueryStatement(String.format("select * from %s", device));
+
+        dataSet =
+            readerPool.executeQueryStatement(
+                String.format("select %s from %s", measurement, device));
+
         DataIterator dataIter = dataSet.iterator();
-        List<String> columnNameList = dataIter.getColumnNameList();
-        List<String> columnTypeList = dataIter.getColumnTypeList();
-        List<MeasurementSchema> schemaList = new ArrayList<>();
-        List<Integer> measureMentValueList = new ArrayList<>();
-        for (int j = 1; j < columnNameList.size(); j++) {
-          String measurement =
-              columnNameList.get(j).substring(columnNameList.get(j).lastIndexOf(".") + 1);
-          if (measurement.contains("`")) {
-            continue;
-          }
-          schemaList.add(
-              new MeasurementSchema(measurement, TSDataType.valueOf(columnTypeList.get(j))));
-          measureMentValueList.add(j + 1);
-        }
-        if (schemaList.isEmpty()) {
-          return null;
-        }
-        tablet = new Tablet(device, schemaList, 300000);
         while (dataIter.next()) {
           int row = tablet.rowSize++;
           tablet.timestamps[row] = dataIter.getLong(1);
-          for (int j = 0; j < schemaList.size(); ++j) {
-            if (dataIter.isNull(measureMentValueList.get(j))) {
-              tablet.addValue(schemaList.get(j).getMeasurementId(), row, null);
-              continue;
-            }
-            switch (schemaList.get(j).getType()) {
-              case BOOLEAN:
-                tablet.addValue(
-                    schemaList.get(j).getMeasurementId(),
-                    row,
-                    dataIter.getBoolean(measureMentValueList.get(j)));
-                break;
-              case INT32:
-                tablet.addValue(
-                    schemaList.get(j).getMeasurementId(),
-                    row,
-                    dataIter.getInt(measureMentValueList.get(j)));
-                break;
-              case INT64:
-                tablet.addValue(
-                    schemaList.get(j).getMeasurementId(),
-                    row,
-                    dataIter.getLong(measureMentValueList.get(j)));
-                break;
-              case FLOAT:
-                tablet.addValue(
-                    schemaList.get(j).getMeasurementId(),
-                    row,
-                    dataIter.getFloat(measureMentValueList.get(j)));
-                break;
-              case DOUBLE:
-                tablet.addValue(
-                    schemaList.get(j).getMeasurementId(),
-                    row,
-                    dataIter.getDouble(measureMentValueList.get(j)));
-                break;
-              case TEXT:
-                tablet.addValue(
-                    schemaList.get(j).getMeasurementId(),
-                    row,
-                    dataIter.getString(measureMentValueList.get(j)));
-                break;
-              default:
-                System.out.println("Migration of this type of data is not supported");
-            }
+          switch (dataType) {
+            case BOOLEAN:
+              ((boolean[]) tablet.values[0])[row] = dataIter.getBoolean(2);
+              break;
+            case INT32:
+              ((int[]) tablet.values[0])[row] = dataIter.getInt(2);
+              break;
+            case INT64:
+              ((long[]) tablet.values[0])[row] = dataIter.getLong(2);
+              break;
+            case FLOAT:
+              ((float[]) tablet.values[0])[row] = dataIter.getFloat(2);
+              break;
+            case DOUBLE:
+              ((double[]) tablet.values[0])[row] = dataIter.getDouble(2);
+              break;
+            case TEXT:
+              ((Binary[]) tablet.values[0])[row] = new Binary(dataIter.getString(2));
+              break;
           }
           if (tablet.rowSize == tablet.getMaxRowNumber()) {
             writerPool.insertTablet(tablet, true);
-            System.out.println("device: " + device + " migrates " + 300000 + " data");
             tablet.reset();
           }
         }
         if (tablet.rowSize != 0) {
           writerPool.insertTablet(tablet);
-          System.out.println("device: " + device + " migrates " + tablet.rowSize + " data");
           tablet.reset();
         }
 
       } catch (Exception e) {
         System.out.println(
-            "Loading the " + i + "-th device: " + device + " failed " + e.getMessage());
+            "Loading the " + i + "-th timeseries: " + series + " failed " + e.getMessage());
         return null;
       } finally {
-        if (dataSet != null) {
-          readerPool.closeResultSet(dataSet);
-        }
-        long endTime = System.nanoTime();
-        long totalTime = endTime - startTime;
-        System.out.println("device ：" + device + " 运行耗时 " + totalTime + " ms");
+        readerPool.closeResultSet(dataSet);
       }
 
-      System.out.println("Loading the " + i + "-th device: " + device + " success");
+      System.out.println("Loading the " + i + "-th timeseries: " + series + " success");
       return null;
     }
   }
